Fix the bug of the program stuck after jobs are completed, by reducing the parameter "mine_num_processes" to 1 in config files (e.g., test_efficiency_1998.json).
It is because >1 values of this param will make each job to alloc >1
processes, and after the job finishes, it will take long long time to
release the CPU resources.

Just pay attention to the code line of "with contextlib.suppress(BrokenPipeError), contextlib.ExitStack() as stack" in jsonql.py. If "mine_num_process" is large, stack will stuck on exiting.

	modified:   cc_net/dedup.py
	modified:   cc_net/execution.py
	modified:   cc_net/jsonql.py
	modified:   cc_net/mine.py
	new file:   config/test_efficiency2.json
	new file:   config/test_efficiency_128.json
	new file:   config/test_efficiency_1998.json
	new file:   config/test_efficiency_64.json
	new file:   killpids.sh
	modified:   run.sh
	new file:   run2.sh
	new file:   run_test.sh
	new file:   size.txt